login page

+++ app/models/mediaManager.php
@@ -10,6 +10,7 @@ require_once '../app/core/DbManager.php';
 
 class MediaManager extends DbManager {
 
+    // ritorna tutte le gifs
     function getAllGifs($user) {
         $stmt = $this->db_connection()->prepare("SELECT  DISTINCT id,title,src,user,owner FROM `favorite` right JOIN gifs on gif_id=id and user=:user_id ");
         $stmt->bindParam(':user_id',$user);
@@ -20,8 +21,9 @@ class MediaManager extends DbManager {
         return $gif;
     }
 
+    //ritorna gifs della categoria $category
     function getCategory($category) {
-        $stmt = $this->db_connection()->prepare("SELECT * FROM `categories`JOIN gifs ON gif_id = id where category = :category");
+        $stmt = $this->db_connection()->prepare("SELECT DISTINCT * FROM categories JOIN gifs ON categories.gif_id = gifs.id LEFT JOIN favorite ON favorite.gif_id=id WHERE category = :category");
         $stmt->bindParam(':category',$category);
         $stmt->execute();
         $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
@@ -29,6 +31,7 @@ class MediaManager extends DbManager {
         return $categories;
     }
 
+    //verifica se $title esiste già come titolo di una gif
     function existTitle($title) {
         $stmt = $this->db_connection()->prepare("SELECT title FROM gifs WHERE title=:title LIMIT 1");
         $stmt->execute([":title" => $title]);
@@ -41,6 +44,7 @@ class MediaManager extends DbManager {
         }
     }
 
+    //ritorna l'utlimo ID inserito per le gif
     function getLastId() {
         $stmt = $this->db_connection()->prepare("SELECT id FROM gifs ORDER BY id DESC LIMIT 1");
         $stmt->execute();
@@ -48,14 +52,15 @@ class MediaManager extends DbManager {
         return $rows;
     }
 
+    //inserisce la gif $gif_id nella categoria $cat
     function uploadCatDb($gif_id,$cat) {
-
         $stmt = $this->db_connection()->prepare("INSERT INTO categories (gif_id, category)  VALUES (:gif_id,:category)");
         $stmt->bindParam(':gif_id',$gif_id);
         $stmt->bindParam(':category',$cat);
         return $stmt->execute();
     }
 
+    //inserisce una nuova gif nel db
     function uploadDb($title,$src,$owner) {
 
         $stmt = $this->db_connection()->prepare("INSERT INTO gifs (title, src,owner)  VALUES (:title,:src,:owner)");
@@ -65,6 +70,7 @@ class MediaManager extends DbManager {
         return $stmt->execute();
     }
 
+    //aggiungeall'utente $user_id la gif favorita $gif_id
     function favGifDb($user_id,$gif_id) {
         $stmt = $this->db_connection()->prepare("INSERT INTO favorite  VALUES (:user_id,:gif_id)");
         $stmt->bindParam(':user_id',$user_id);
@@ -73,12 +79,15 @@ class MediaManager extends DbManager {
     }
 
 
-    function delFavDb($gif_id) {
-        $stmt = $this->db_connection()->prepare("DELETE FROM favorite WHERE gif_id=:gif_id");
+    //cancella la gif favorita  $gif_id
+    function delFavDb($gif_id,$user_id) {
+        $stmt = $this->db_connection()->prepare("DELETE FROM favorite WHERE gif_id=:gif_id AND user=:user_id");
         $stmt->bindParam(':gif_id',$gif_id);
+        $stmt->bindParam(':user_id',$user_id);
         return $stmt->execute();
     }
 
+    //ritorna tutte le fig dell'artista $user_id
     function getArtistGif($user_id) {
         $stmt = $this->db_connection()->prepare("SELECT * FROM gifs   WHERE owner=:owner");
         $stmt->bindParam(':owner',$user_id);
@@ -89,8 +98,9 @@ class MediaManager extends DbManager {
         return $gif;
     }
 
+    //ritorna tuttie le gif favorite dell utente $user_id
     function getArtistFav($user_id) {
-        $stmt = $this->db_connection()->prepare(" SELECT DISTINCT id,title,src,user,owner FROM `favorite`JOIN gifs on gif_id=id and user=:user_id");
+        $stmt = $this->db_connection()->prepare(" SELECT DISTINCT id,title,src,user,owner FROM `favorite`JOIN gifs ON gif_id=id AND user=:user_id");
         $stmt->bindParam(':user_id',$user_id);
         $stmt->execute();
         $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
@@ -99,12 +109,14 @@ class MediaManager extends DbManager {
         return $gif;
     }
 
+    //elimina la gif $gif_id
     function delUploadsDb($gif_id) {
         $stmt = $this->db_connection()->prepare(" DELETE FROM gifs WHERE id=:gif_id");
         $stmt->bindParam(':gif_id',$gif_id);
         return $stmt->execute();
     }
 
+    //ritorna tuttii gli utenti/artisti
     function getAllArtistsDb() {
         $stmt = $this->db_connection()->prepare("SELECT nickname FROM users");
         $stmt->execute();
@@ -113,5 +125,4 @@ class MediaManager extends DbManager {
         $artists['art'] = $rows;
         return $artists;
     }
-
 }
